Convert into an internal API
